+++ src/franz/openrdf/vocabulary/__init__.py
@@ -22,29 +22,4 @@ from .rdf import RDF
 from .rdfs import RDFS
 from .xmlschema import XMLSchema
 
-__all__ = ['OWL', 'RDF', 'RDFS', 'XMLSchema', 'canonical_uri_object']
-
-_uri_map = {}
-
-
-def _compute_uri_map():
-    for cls in OWL, RDF, RDFS, XMLSchema:
-        for name, uri in iteritems(cls.__dict__):
-            if name.isupper and isinstance(uri, URI):
-                _uri_map[uri.uri] = uri
-
-_compute_uri_map()
-
-
-def canonical_uri_object(uri):
-    """
-    Check if there is a predefined constant for ``uri`` and return it.
-
-     If there is no such consatnt ``None`` is returned.
-
-    :param uri: URI string.
-    :type uri: string
-    :return: An URI object or ``None``.
-    :rtype: URI
-    """
-    return _uri_map.get(uri)
+__all__ = ['OWL', 'RDF', 'RDFS', 'XMLSchema']