+++ CHANGES.rst
@@ -10,7 +10,7 @@ User data access
 
 AllegroGraph allows each user to store arbitrary key-value data
 on the server. This storage can now be accessed from Python by
-using new ``AllegroGraphServer` methods:
+using new ``AllegroGraphServer`` methods:
 
   - ``listUserData()``
   - ``getUserData()``